+++ cmd/instability.go
@@ -6,6 +6,7 @@ import (
 	"github.com/fdaines/spm-go/model"
 	"github.com/fdaines/spm-go/utils"
 	"github.com/fdaines/spm-go/utils/output"
+	pkg "github.com/fdaines/spm-go/utils/packages"
 	"github.com/spf13/cobra"
 )
 
@@ -25,13 +26,13 @@ func init() {
 func analyzeInstability(cmd *cobra.Command, args []string) {
 	utils.ExecuteWithTimer(func() {
 		utils.PrintMessage("Instability analysis started.")
-		mainPackage, err := getMainPackage()
+		mainPackage, err := pkg.GetMainPackage()
 		if err != nil {
 			fmt.Printf("Error: %+v\n", err)
 			return
 		}
 		var afferentMap = make(map[string][]string)
-		pkgsInfo := getBasicPackagesInfo()
+		pkgsInfo := pkg.GetBasicPackagesInfo()
 		utils.PrintMessage("Gathering package metrics, please wait until the command is finished running.")
 		for index, pkgInfo := range pkgsInfo {
 			utils.PrintStep()
